add user and profile interface and rewrite concrete classes
