+++ src/edu/saddleback/cs4b/Backend/Server/ClientCommunication.java
@@ -6,7 +6,6 @@ import edu.saddleback.cs4b.Backend.Utilitys.Profile;
 import java.io.*;
 import java.net.Socket;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 /**
